add testing on previous tasks

+++ models/hypernet/helper.py
@@ -25,7 +25,7 @@ def base_train(model, trainloader, optimizer, scheduler, epoch, args):
         logits, feature = model(data, encoded=True)
         logits_ = logits[:, :args.base_class]
     
-        loss = F.cross_entropy(logits, train_label)
+        loss = F.cross_entropy(logits, train_label.long())
 
         acc = count_acc(logits_, train_label)     
         total_loss = loss
@@ -55,7 +55,7 @@ def test(model, testloader, epoch,args, session,validation=True):
             data, test_label = [_.to(args.device) for _ in batch]
             logits = model(data)
             logits = logits[:, :test_class]
-            loss = F.cross_entropy(logits, test_label)
+            loss = F.cross_entropy(logits, test_label.long())
             acc = count_acc(logits, test_label)
             vl.add(loss.item())
             va.add(acc)
